requirements.txt
Django==5.1.3
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.0
flower==2.0.1
djangorestframework-simplejwt==5.3.1
